+++ packages/platform_playlist_import/lib/platform_playlist_import.dart
@@ -1,5 +1,6 @@
 library;
 
+export "src/indexed_m3u_playlist.dart";
 export "src/pipeline/large_playlist_worker_models.dart";
 export "src/m3u_playlist_parser.dart";
 export "src/m3u_parser_service.dart";